requirements.txt
aiohttp>=3.9.0
aiohttp-cors>=0.7.0
python-dotenv>=1.0.0
pyyaml>=6.0
colorama>=0.4.6
rich>=13.0.0
click>=8.1.0
aiosqlite>=0.19.0
jinja2>=3.1.0
cryptography>=41.0.0
